ai-tinkerers-new-york-march-27-2024/pyproject.toml
[tool.poetry]
name = "ai-tinkerers-new-york-march-27-2024"
version = "0.0.0"
description = "Demo of Mirascope Tools and Extraction for the AI Tinkerers New York Event on March 27, 2024"
license = "MIT"
authors = ["William Bakst <william@mirascope.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
mirascope = { version = "^0.7.1", extras = ["anthropic"] }
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
ruff = "^0.1.5"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 88
target-version = "py38"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
exclude = ["venv"]
